# docs/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Cerefox /api/v1",
    "version": "1.1.0",
    "description": "The HTTP API served by `cerefox web`. Intended for callers that embed Cerefox rather than\nusing MCP or the CLI — MCP remains the recommended path for AI agents.\n\nGENERATED by `bun scripts/gen_openapi.ts`; do not edit by hand. Paths and every input a route\nreads come from the route handlers, summaries from the endpoint table in docs/guides/api.md,\nand body shapes from the zod schemas in `_shared/schemas/` (responses verified against a live\nserver by `api-schema-truth.test.ts`).\n\nAUTH: a request arriving on loopback needs no credential. From any other interface, an\n`X-API-Key` (or `Authorization: Bearer`) matching `CEREFOX_API_KEY` is required.\nSee docs/guides/securing-local-access.md.\n\nIDENTITY: send `X-Cerefox-Author` (and optionally `X-Cerefox-Author-Type`) to be recorded as\nyourself. Without them you are recorded as the bundled web UI.\n\nCONCURRENCY: read before you write. A content update (`POST /ingest` with `document_id` or\n`update_if_exists`, `POST /documents/{id}/edit`, `/upload`) needs the `content_hash` you read\nas `expected_content_hash`, or an explicit `last_write_wins`. `DELETE /documents/{id}` needs it\ntoo when you identify yourself, as `X-Cerefox-Expected-Content-Hash` or a query parameter.\nNeither is a 400 (`CEREFOX_TOKEN_REQUIRED`); a stale hash is a 409 carrying `current_hash`.\nRestore and purge take no hash."
  },
  "servers": [
...
  "security": [
    {},
    {
      "apiKey": []
    },
    {
      "bearerAuth": []
    }
  ],
  "components": {
    "schemas": {
      "Error": {
        "type": "object",
        "description": "Error body. `detail` is the human-readable reason. Write routes also return `success: false`; the ingest and edit routes also carry the same text as `error` (and edit as `message`), and a concurrency conflict adds `current_hash`.",
        "properties": {
          "detail": {
            "type": "string"
          },
          "success": {
            "type": "boolean",
            "enum": [
              false
            ]
          },
          "error": {
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "current_hash": {
            "type": "string",
            "description": "On a 409 conflict: the document's hash now. Re-read, merge, retry with it."
          }
        },
        "additionalProperties": true
      }
    },
    "securitySchemes": {
      "apiKey": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key"
      },
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "The same `CEREFOX_API_KEY`, as a bearer token."
      }
    }
  },
  "paths": {
...
    "/api/v1/documents/{document_id}/upload": {
      "post": {
        "summary": "`expected_content_hash` form field",
        "operationId": "post_documents_document_id_upload",
        "parameters": [
          {
            "name": "document_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "X-Cerefox-Author",
            "in": "header",
            "required": false,
            "description": "Who is calling. Omit all three identity headers to be recorded as the bundled web UI (`web-ui`, access path `webapp`); supply any to be recorded as yourself with access path `api`.",
            "schema": {
              "type": "string",
              "minLength": 1
            }
          },
          {
            "name": "X-Cerefox-Requestor",
            "in": "header",
            "required": false,
            "description": "Recorded in the usage log. Defaults to `X-Cerefox-Author`.",
            "schema": {
              "type": "string",
              "minLength": 1
            }
          },
          {
            "name": "X-Cerefox-Author-Type",
            "in": "header",
            "required": false,
            "description": "`agent` makes a new document land in `pending_review`, as it does over MCP.",
            "schema": {
              "type": "string",
              "enum": [
                "user",
                "agent"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "document_id": {
                      "type": "string"
                    },
                    "title": {
                      "type": "string"
                    },
                    "updated": {
                      "type": "boolean"
                    }
                  },
                  "required": [
                    "success",
                    "document_id",
                    "title",
                    "updated"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "400": {
            "description": "The request is malformed or incomplete: a bad parameter or body, an unknown body field, a blank identity header, or a content update with no `expected_content_hash` and no `last_write_wins` (`CEREFOX_TOKEN_REQUIRED`).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "An API key is required: a non-loopback request while `CEREFOX_API_KEY` is set. See docs/guides/securing-local-access.md.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "No such document, version or project; or the feature is off on this store.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "Conflict with the current state: a stale `expected_content_hash` (`CEREFOX_CONFLICT`; the body carries `current_hash` — re-read, merge, retry; never resolve by overwriting blindly), a document in the trash, or a duplicate name.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "422": {
            "description": "Well-formed, but fails a semantic check: content linking to document ids that do not exist, or a missing required parameter.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Unhandled server or database error.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "503": {
            "description": "A dependency is unavailable: no embedder configured, or the deployed schema needs `cerefox server deploy`.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "required": [
                  "file"
                ],
                "properties": {
                  "file": {
                    "type": "string",
                    "format": "binary",
                    "description": "The replacement content, converted to markdown."
                  },
                  "expected_content_hash": {
                    "type": "string",
                    "description": "The content_hash you read the document at. Required on a content update unless last_write_wins is `true`."
                  },
                  "last_write_wins": {
                    "type": "string",
                    "enum": [
                      "true",
                      "false"
                    ],
                    "description": "Skip the concurrency check."
                  },
                  "author": {
                    "type": "string"
                  },
                  "requestor": {
                    "type": "string"
                  },
                  "author_type": {
                    "type": "string",
                    "enum": [
                      "user",
                      "agent"
                    ]
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/ingest/file": {
      "post": {
        "summary": "`expected_content_hash` form field",
        "operationId": "post_ingest_file",
        "parameters": [
          {
            "name": "X-Cerefox-Author",
            "in": "header",
            "required": false,
            "description": "Who is calling. Omit all three identity headers to be recorded as the bundled web UI (`web-ui`, access path `webapp`); supply any to be recorded as yourself with access path `api`.",
            "schema": {
              "type": "string",
              "minLength": 1
            }
          },
          {
            "name": "X-Cerefox-Requestor",
            "in": "header",
            "required": false,
            "description": "Recorded in the usage log. Defaults to `X-Cerefox-Author`.",
            "schema": {
              "type": "string",
              "minLength": 1
            }
          },
          {
            "name": "X-Cerefox-Author-Type",
            "in": "header",
            "required": false,
            "description": "`agent` makes a new document land in `pending_review`, as it does over MCP.",
            "schema": {
              "type": "string",
              "enum": [
                "user",
                "agent"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "document_id": {
                      "type": "string"
                    },
                    "title": {
                      "type": "string"
                    },
                    "skipped": {
                      "type": "boolean"
                    },
                    "updated": {
                      "type": "boolean"
                    },
                    "note": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "success",
                    "document_id",
                    "title",
                    "skipped",
                    "updated"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "400": {
            "description": "The request is malformed or incomplete: a bad parameter or body, an unknown body field, a blank identity header, or a content update with no `expected_content_hash` and no `last_write_wins` (`CEREFOX_TOKEN_REQUIRED`).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "An API key is required: a non-loopback request while `CEREFOX_API_KEY` is set. See docs/guides/securing-local-access.md.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "No such document, version or project; or the feature is off on this store.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "Conflict with the current state: a stale `expected_content_hash` (`CEREFOX_CONFLICT`; the body carries `current_hash` — re-read, merge, retry; never resolve by overwriting blindly), a document in the trash, or a duplicate name.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "422": {
            "description": "Well-formed, but fails a semantic check: content linking to document ids that do not exist, or a missing required parameter.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Unhandled server or database error.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "503": {
            "description": "A dependency is unavailable: no embedder configured, or the deployed schema needs `cerefox server deploy`.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "required": [
                  "file"
                ],
                "properties": {
                  "file": {
                    "type": "string",
                    "format": "binary",
                    "description": "The file to convert to markdown and ingest (.md, .txt, .pdf, .docx, …)."
                  },
                  "title": {
                    "type": "string",
                    "description": "Defaults to the filename."
                  },
                  "update_existing": {
                    "type": "string",
                    "enum": [
                      "true",
                      "false"
                    ],
                    "description": "Update the document matching this file's path or title instead of creating another."
                  },
                  "project_ids": {
                    "type": "string",
                    "description": "Comma-separated project ids."
                  },
                  "metadata": {
                    "type": "string",
                    "description": "A JSON-encoded object of metadata key/value pairs."
                  },
                  "expected_content_hash": {
                    "type": "string",
                    "description": "The content_hash you read the document at. Required on a content update unless last_write_wins is `true`."
                  },
                  "last_write_wins": {
                    "type": "string",
                    "enum": [
                      "true",
                      "false"
                    ],
                    "description": "Skip the concurrency check."
                  },
                  "author": {
                    "type": "string"
                  },
                  "requestor": {
                    "type": "string"
                  },
                  "author_type": {
                    "type": "string",
                    "enum": [
                      "user",
                      "agent"
                    ]
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}