# packages/memory/src/web/routes/ingest.ts
/**
 * Ingestion endpoints (Part 25F — swap from 503 stubs to real handlers):
 *
 *   POST /api/v1/ingest                            — paste mode
 *   POST /api/v1/ingest/file                       — multipart file upload
 *   POST /api/v1/documents/{document_id}/upload    — replace existing content
 *
 * v0.6 (Part 24H) shipped these as 503 stubs with a friendly toast.
 * v0.7 (this Part) wires them to the in-process `IngestionPipeline`.
 * Frontend's `V07IngestionDeferredError` toast detector stays in
 * `api/client.ts` as dead code — no frontend changes required.
 *
 * Python source (kept through v0.7.x for the legacy Python web):
 *   - api_ingest_paste     (routes_api.py:1030)
 *   - api_ingest_file      (routes_api.py:1074)
 *   - api_upload_content   (routes_api.py:994)
 *
 * Embedder dependency: when `ctx.openAiApiKey` is null these endpoints
 * return 503 with `error: "Embedder not available"` — matches Python's
 * `Embedder not available` 503 shape.
 */

import { resolveEmbedderKind } from "../../../../../_shared/embeddings/index.ts";
import { Hono } from "hono";

import { fileToMarkdown } from "../../ingestion/file-to-markdown.ts";
import { IngestionPipeline } from "../../ingestion/pipeline.ts";
import {
  ConcurrencyConflictError,
  ConcurrencyTokenRequiredError,
} from "../../ingestion/types.ts";
import type { WebContext } from "../context.ts";
import { logWebUsage } from "../usage.ts";
import { resolveCallerIdentity } from "../identity.ts";

interface IngestResponse {
  success: boolean;
  document_id?: string;
  title?: string;
  skipped?: boolean;
  updated?: boolean;
  note?: string;
  error?: string;
  /** Present on a 409 so the caller can re-read without a second round trip —
   *  the same field `documents-write.ts` returns on an edit conflict. */
  current_hash?: string;
  /** Same text as `error`. Present because the rest of `/api/v1` puts its
   *  human-readable reason in `detail`, and the frontend's `ApiError` reads
   *  exactly that field to build its message (`api/client.ts`). Emitting only
   *  `error` on a non-2xx would show the user a bare "API error 400". */
  detail?: string;
}

function notReady(error: string): IngestResponse {
  return { success: false, error, detail: error };
}

/**
 * Map a pipeline failure to its HTTP status (#232).
 *
 * These routes used to answer `200 {success:false, error}` for EVERY pipeline
 * failure, including a refused write. That was survivable while the only
 * caller was the bundled UI, which reads `success`. #226 opened the surface to
 * other clients, and an HTTP client that checks `resp.ok` — what
 * `raise_for_status()`, `curl -f` and every retry wrapper do by default —
 * would read a refused write as a successful one. A concurrency refusal is the
 * worst case to miss, because the caller's next move depends on knowing.
 *
 * The mapping is not invented here: `documents-write.ts` has mapped these same
 * two typed errors to 400/409 since iter-32. This brings the three ingest
 * routes onto that existing convention rather than adding a second one.
 *
 * `success: false` and `error` stay in the body, so a client reading either of
 * them keeps working; only the status (and the added `detail`) are new.
 */
function ingestFailure(err: unknown): [IngestResponse, 400 | 404 | 409 | 422 | 500] {
  if (err instanceof ConcurrencyConflictError) {
    return [
      {
        ...notReady(err.message),
        ...(err.currentHash ? { current_hash: err.currentHash } : {}),
      },
      409,
    ];
  }
  if (err instanceof ConcurrencyTokenRequiredError) {
    return [notReady(err.message), 400];
  }
  const msg = err instanceof Error ? err.message : String(err);
  // Caller-state failures are not server faults (#296). Classified by the
  // pipeline's own wording, the same way `documents-write.ts` classifies the
  // edit path: an update by an unknown id, an update of a trashed document,
  // and content linking to ids that do not exist.
  if (/^Document (.* )?not found/.test(msg)) return [notReady(msg), 404];
  if (msg.includes("soft-deleted")) return [notReady(msg), 409];
  if (msg.includes("CEREFOX_UNRESOLVED_LINKS")) return [notReady(msg), 422];
  return [notReady(msg), 500];
}

/** Every top-level field `POST /ingest` reads. `mode` is sent by the bundled
 *  UI and carries no meaning; the identity fields are read by
 *  `resolveCallerIdentity`. `ingest-contract.test.ts` pins this list to the
 *  published request schema in both directions. */
export const INGEST_FIELDS = new Set([
  "title",
  "content",
  "document_id",
  "expected_content_hash",
  "last_write_wins",
  "update_existing",
  "update_if_exists",
  "project_ids",
  "project_names",
  "project_name",
  "metadata",
  "source",
  "mode",
  "author",
  "requestor",
  "author_type",
]);

/** Type checks a cast would otherwise hide until the RPC or a jsonb column. */
function ingestShapeError(body: Record<string, unknown>): string | null {
  const isStrArray = (v: unknown) => Array.isArray(v) && v.every((x) => typeof x === "string");
  if (body.metadata != null && (typeof body.metadata !== "object" || Array.isArray(body.metadata))) {
    return "metadata must be a JSON object of key/value pairs.";
  }
  if (body.project_ids != null && !isStrArray(body.project_ids)) return "project_ids must be an array of strings.";
  if (body.project_names != null && !isStrArray(body.project_names)) return "project_names must be an array of strings.";
  for (const k of ["project_name", "source", "document_id", "expected_content_hash"]) {
    if (body[k] != null && typeof body[k] !== "string") return `${k} must be a string.`;
  }
  for (const k of ["last_write_wins", "update_existing", "update_if_exists"]) {
    if (body[k] != null && typeof body[k] !== "boolean") return `${k} must be a boolean.`;
  }
  return null;
}

export function registerIngestRoutes(app: Hono, ctx: WebContext): void {
  // ── POST /api/v1/ingest (paste) ───────────────────────────────────────────
  app.post("/api/v1/ingest", async (c) => {
    if (!ctx.openAiApiKey && resolveEmbedderKind() !== "local") {
      return c.json(notReady("Embedder not available"), 503);
    }

    let body: Record<string, unknown>;
    try {
      body = (await c.req.json()) as Record<string, unknown>;
    } catch {
      return c.json(notReady("Invalid JSON body"), 400);
    }
    // Unknown fields are refused, not dropped (#296). The published contract
    // once listed six fields this route ignored; a caller that sent
    // `update_if_exists` got a duplicate document instead of an update, and
    // nothing told it why. A typo now costs a 400 rather than a silent miswrite.
    const unknown = Object.keys(body).filter((k) => !INGEST_FIELDS.has(k));
    if (unknown.length > 0) {
      return c.json(
        notReady(`Unknown field(s): ${unknown.join(", ")}. Accepted: ${[...INGEST_FIELDS].join(", ")}.`),
        400,
      );
    }
    const title = String(body.title ?? "").trim();
    const content = String(body.content ?? "");

    // Real statuses (#296): these answered 200 {success:false}, the pattern
    // #232 removed from every other refusal on this route.
    if (!title) return c.json(notReady("Title is required."), 400);
    if (!content.trim()) return c.json(notReady("Content cannot be empty."), 400);
    const shapeError = ingestShapeError(body);
    if (shapeError) return c.json(notReady(shapeError), 400);

    const who = resolveCallerIdentity(c, body);
    if (!who.ok) return c.json({ detail: who.detail }, 400);

    try {
      const pipeline = new IngestionPipeline({
        supabase: ctx.supabase,
        openAiApiKey: ctx.openAiApiKey ?? "",
      });
      const result = await pipeline.ingestText({
        text: content.trim(),
        title,
        source: typeof body.source === "string" && body.source.trim() ? body.source.trim() : "paste",
        projectIds: Array.isArray(body.project_ids) ? (body.project_ids as string[]) : null,
        projectNames: Array.isArray(body.project_names) ? (body.project_names as string[]) : null,
        projectName: typeof body.project_name === "string" ? body.project_name : null,
        metadata: (body.metadata as Record<string, unknown> | null | undefined) ?? null,
        // `update_if_exists` is the name MCP and the CLI use; `update_existing`
        // is the one the bundled UI has always sent. Either means the same.
        updateExisting: body.update_existing === true || body.update_if_exists === true,
        documentId: (body.document_id as string | undefined) ?? null,
        author: who.identity.author,
        authorType: who.identity.authorType,
        // The concurrency contract (#296). Never read before v1.15.2, so an
        // update by document_id was refused with CEREFOX_TOKEN_REQUIRED
        // whatever the caller sent.
        expectedContentHash:
          typeof body.expected_content_hash === "string" && body.expected_content_hash.trim() !== ""
            ? body.expected_content_hash.trim()
            : null,
        lastWriteWins: body.last_write_wins === true,
      });
      const skipped = result.action === "skipped";
      const resp: IngestResponse = {
        success: !skipped,
        document_id: result.documentId,
        title: result.title,
        skipped,
        updated: result.reindexed,
      };
      if (result.note) resp.note = result.note;
      logWebUsage(ctx, {
        operation: "ingest",
        document_id: result.documentId,
        requestor: who.identity.requestor,
        access_path: who.identity.accessPath,
      });
      return c.json(resp, 200);
    } catch (err) {
      const [body, status] = ingestFailure(err);
      return c.json(body, status);
    }
  });

  // ── POST /api/v1/ingest/file (multipart) ──────────────────────────────────
  app.post("/api/v1/ingest/file", async (c) => {
    if (!ctx.openAiApiKey && resolveEmbedderKind() !== "local") {
      return c.json(notReady("Embedder not available"), 503);
    }

    let form: { [k: string]: unknown };
    try {
      form = await c.req.parseBody();
    } catch (err) {
      return c.json(
        notReady(`Invalid multipart body: ${err instanceof Error ? err.message : err}`),
        400,
      );
    }
    const file = form.file as File | undefined;
    if (!file) return c.json(notReady("file field is required."), 400);

    let text: string;
    try {
      text = await fileToMarkdown(file.name, Buffer.from(await file.arrayBuffer()));
    } catch (err) {
      return c.json(notReady(err instanceof Error ? err.message : String(err)), 400);
    }
    const titleField = String(form.title ?? "").trim();
    const docTitle = titleField || file.name || "Untitled";
    const updateExisting = String(form.update_existing ?? "false") === "true";
    const projectIdsStr = String(form.project_ids ?? "");
    const projectIds = projectIdsStr
      ? projectIdsStr.split(",").map((s) => s.trim()).filter(Boolean)
      : null;
    const metadataStr = String(form.metadata ?? "");
    let metadata: Record<string, string> | null = null;
    if (metadataStr) {
      try {
        metadata = JSON.parse(metadataStr) as Record<string, string>;
      } catch {
        metadata = null;
      }
    }

    // Multipart form fields are strings, so the same resolver reads them.
    const who = resolveCallerIdentity(c, form);
    if (!who.ok) return c.json({ detail: who.detail }, 400);

    try {
      const pipeline = new IngestionPipeline({
        supabase: ctx.supabase,
        openAiApiKey: ctx.openAiApiKey ?? "",
      });
      const result = await pipeline.ingestText({
        text,
        title: docTitle,
        source: "file",
        sourcePath: file.name,
        projectIds,
        metadata,
        updateExisting,
        author: who.identity.author,
        authorType: who.identity.authorType,
        // Same contract as POST /ingest and /upload (#296): an update_existing
        // that matches a document is a content update, and needs the hash or an
        // explicit last_write_wins. Without these the route had no way to say
        // either, so matching an existing document was always refused.
        expectedContentHash:
          typeof form.expected_content_hash === "string" && form.expected_content_hash.trim() !== ""
            ? form.expected_content_hash.trim()
            : null,
        lastWriteWins: String(form.last_write_wins ?? "") === "true",
      });
      const skipped = result.action === "skipped";
      logWebUsage(ctx, {
        operation: "ingest",
        document_id: result.documentId,
        requestor: who.identity.requestor,
        access_path: who.identity.accessPath,
      });
      return c.json(
        {
          success: !skipped,
          document_id: result.documentId,
          title: result.title,
          skipped,
          updated: result.reindexed,
          // The WHY on a skip (e.g. "identical content is in the TRASH —
          // restore it"). The paste branch always forwarded it; dropping it
          // here left file re-uploads with a bare unexplained "skipped".
          ...(result.note ? { note: result.note } : {}),
        },
        200,
      );
    } catch (err) {
      const [body, status] = ingestFailure(err);
      return c.json(body, status);
    }
  });

  // ── POST /api/v1/documents/{id}/upload (replace existing) ────────────────
  app.post("/api/v1/documents/:document_id/upload", async (c) => {
    if (!ctx.openAiApiKey && resolveEmbedderKind() !== "local") {
      return c.json(notReady("Embedder not available"), 503);
    }
    const documentId = c.req.param("document_id");

    let form: { [k: string]: unknown };
    try {
      form = await c.req.parseBody();
    } catch (err) {
      return c.json(
        notReady(`Invalid multipart body: ${err instanceof Error ? err.message : err}`),
        400,
      );
    }
    const file = form.file as File | undefined;
    if (!file) return c.json(notReady("file field is required."), 400);

    let text: string;
    try {
      text = await fileToMarkdown(file.name, Buffer.from(await file.arrayBuffer()));
    } catch (err) {
      return c.json(notReady(err instanceof Error ? err.message : String(err)), 400);
    }

    // Fetch the existing doc so we can preserve its title if the upload
    // doesn't provide one (matches Python: keeps existing.title or
    // falls back to filename).
    const { data: existing } = await ctx.supabase
      .from("cerefox_documents")
      .select("id, title")
      .eq("id", documentId)
      .maybeSingle();
    if (!existing) return c.json(notReady("Document not found"), 404);

    const title =
      (existing.title as string | null) || file.name || "Untitled";

    const who = resolveCallerIdentity(c, form);
    if (!who.ok) return c.json({ detail: who.detail }, 400);

    try {
      const pipeline = new IngestionPipeline({
        supabase: ctx.supabase,
        openAiApiKey: ctx.openAiApiKey ?? "",
      });
      const result = await pipeline.updateDocument({
        documentId,
        text,
        title,
        source: "file",
        author: who.identity.author,
        authorType: who.identity.authorType,
        // Optimistic concurrency (iter-40, #228). This route passed NEITHER a
        // hash nor last_write_wins, so from v0.11.0 — when content updates
        // began requiring one — every call returned CEREFOX_TOKEN_REQUIRED.
        // It went unnoticed for eleven releases because the only test covering
        // it lives in the web-integration suite, which had been skipping since
        // v0.9.0 on a renamed probe verb.
        //
        // It takes the SAME contract as every other content update: send the
        // hash you read, or say last_write_wins explicitly. There is no
        // implicit last-write-wins default, because there is no working caller
        // to stay compatible with — the endpoint has been a hard error since
        // v0.11.0, so the strict semantics cost nothing and a silent clobber
        // would have been a new way to lose data on a surface we are opening
        // to more clients.
        expectedContentHash:
          typeof form.expected_content_hash === "string" &&
          form.expected_content_hash.trim() !== ""
            ? form.expected_content_hash.trim()
            : null,
        lastWriteWins: String(form.last_write_wins ?? "") === "true",
      });
      logWebUsage(ctx, {
        operation: "ingest",
        document_id: result.documentId,
        requestor: who.identity.requestor,
        access_path: who.identity.accessPath,
      });
      return c.json(
        {
          success: true,
          document_id: result.documentId,
          title: result.title,
          updated: result.reindexed,
        },
        200,
      );
    } catch (err) {
      const [body, status] = ingestFailure(err);
      return c.json(body, status);
    }
  });
}